Allow services to specify SSL certificates
Add command line option to specify locations for the default SSL
certificate, and rename the stats SSL certificate options, so that we
can output them to our HAproxy configuration. Also provide ability to
annotate services with locations for the SSL certificates so they can be
used for SSL termination. Finally, update make_secret.go to use a
configuration file input so that it will generate our secrets volume as
specified.